Add elementary support for minimal-pull under Pathom 3 (#28)
Limitations: It only compares what the DB can offer against what the user _directly_ asked for. If a resolver needs extra input to compute a derived attribute, these will not be included.

Essentially the same as #27 but preserves `prune-query` with the original signature so we do not break any client code that can be relying on it.